matched func_8005B548_5C148 6 attempts

+++ include/3E160.h
@@ -8,4 +8,4 @@ void initRace(void);
 void func_80057ABC_586BC(s32 arg0, s32 arg1);
 void func_800B4B30_1E1BE0(s16 arg0, s16 arg1, s16 arg2, s16 arg3);
 void func_80037E78_38A78(Entry1 *arg0);
-Player *func_8005B548_5C148(void *arg0, s16 arg1, s32 arg2);
+Player *func_8005B548_5C148(void *arg0, s32 arg1, s32 arg2);
